Implementation of usage of behaviour

+++ src/main/java/lasersharks/behaviour/checkforloss/DefaultCheckForLossBehaviour.java
@@ -9,7 +9,7 @@ import lasersharks.behaviour.interfaces.CheckForLossBehaviour;
  */
 public class DefaultCheckForLossBehaviour implements CheckForLossBehaviour {
   @Override
-  public void onCollisionPlayerLoses(int size) {
+  public void onCollisionPlayerLoses(float size) {
   }
 
 }
