+++ src/main/java/lasersharks/behaviour/interfaces/CheckForLossBehaviour.java
@@ -1,5 +1,5 @@
 package lasersharks.behaviour.interfaces;
 
 public interface CheckForLossBehaviour {
-  void onCollisionPlayerLoses(int size);
+  void onCollisionPlayerLoses(float size);
 }
